% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract-network.R
\name{extractNetwork}
\alias{extractNetwork}
\title{Extract collaborators network from Google Scholar page}
\usage{
extractNetwork(id, n = 500, largest_component = TRUE, ...)
}
\arguments{
\item{id}{Character string specifying the Google Scholar ID.}

\item{n}{Maximum number of publications to retrieve.}

\item{largest_component}{If \code{TRUE}, keep only largest component in network}

\item{...}{Other options to pass to \code{get_publications} function}
}
\description{
Uses \code{scholar} package to scrape Google Scholar page of an author
(determined by ID) and returns a list with a list of edges and a data frame
with node-level information
}
\details{
extractNetwork
}
\examples{
\dontrun{
## Download Google Scholar network data for a sample user
d <- extractNetwork(id="jGLKJUoAAAAJ", n=500)
## Plot network into file called \\code{network.html}
plotNetwork(d$nodes, d$edges, file="network.html")
}
}

