% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-network.R
\name{plotNetwork}
\alias{plotNetwork}
\title{Plot collaborators network from Google Scholar page}
\usage{
plotNetwork(nodes, edges, file = "network.html", width = 550,
  height = 400, opacity = 0.75, fontsize = 10, charge = -500, ...)
}
\arguments{
\item{nodes}{Data frame with node information returned by \code{extractNetwork}.}

\item{edges}{Data frame with edge list returned by \code{extractNetwork}.}

\item{file}{File where network visualization will be exported to.}

\item{width}{numeric width for the network graph's frame area in pixels}

\item{height}{numeric height for the network graph's frame area in pixels.}

\item{opacity}{numeric value of the proportion opaque you would like the graph elements to be.}

\item{fontsize}{numeric font size in pixels for the node text labels.}

\item{charge}{numeric value indicating either the strength of the node repulsion (negative value) or attraction (positive value).}

\item{...}{Other options to pass to \code{networkD3} function

#'}
}
\description{
Takes value from \code{extractNetwork} function and visualizes network
using networkD3.
}
\details{
plotNetwork
}
\examples{
\dontrun{
## Download Google Scholar network data for a sample user
d <- extractNetwork(id="jGLKJUoAAAAJ", n=500)
## Plot network into file called \\code{network.html}
plotNetwork(d$nodes, d$edges, file="network.html")
}
}

