% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scholarnetwork-package.R
\docType{package}
\name{scholarnetwork-package}
\alias{scholarnetwork}
\alias{scholarnetwork-package}
\title{Extract and Visualize Google Scholar Collaboration Networks}
\description{
This package provides functions to extracts publication information from
Google Scholar, create network of collaborators based on co-authored projects,
and visualize these networks using a force-directed layout algorithm.
}
\author{
Pablo Barbera \email{pablo.barbera@nyu.edu}
}
\seealso{
\code{\link{extractNetwork}}, \code{\link{plotNetwork}}
}

